ENH: Call InitializeObjectBase() in New() methods
In VTK8 it's necessary for New() methods to call InitializeObjectBase() on the
new object for proper tracking with vtkDebugLeaks. The standard macros
(vtkStandardNewMacro, vtkObjectFactoryNewMacro) handle this. For those classes
that don't use the macros, add calls to InitializeObjectBase().

Continue to support earlier versions of VTK by wrapping calls to
InitializeObjectBase() in checks for the preprocessor define
VTK_HAS_INITIALIZE_OBJECT_BASE.

It's possible that many of these implementations could be simplified by using
the standard macros.

See also:
- https://github.com/Kitware/VTK/commit/e5c793dbdf87e838bb2b60c6a5905ced0e5548f9
- http://public.kitware.com/pipermail/vtk-developers/2016-September/034332.html